requirements.txt
click==7.1.2
dnspython==2.0.0
email-validator==1.1.2
flake8==3.8.4
Flask==1.1.2
Flask-Login==0.5.0
Flask-WTF==0.14.3
gunicorn==20.0.4
idna==2.10
importlib-metadata==3.3.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.6
pycodestyle==2.6.0
pyflakes==2.2.0
python-decouple==3.4
python-dotenv==0.15.0
typing-extensions==3.7.4.3
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.4.0
